+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `defaultValue` to `files` form element validation
+
 ## [0.9.15] - 2021-06-23
 
 ### Added